refactoring client functions for fence (#864)
* refactoring client functions for fence

* fixing

* error when calling tasks function in props

* fix travis

* fix the expect issue

* fixing the createClient

* fixing travis again
