+++ services/apis/fence/fenceProps.js
@@ -21,18 +21,15 @@ function createClient(clientName, userName, clientType, arboristPolicies = null)
 
   // see test_setup.js for the ARBORIST_... global flag setup
   if (arboristPolicies && process.env.ARBORIST_CLIENT_POLICIES) {
-    fenceCmd = `${fenceCmd} --arborist http://arborist-service/`;
+    fenceCmd = `${fenceCmd} --arborist http://arborist-service/ --policies ${arboristPolicies}`;
   }
 
-  fenceCmd = `${fenceCmd} client-create --client ${clientName} --user ${userName} --urls https://${process.env.HOSTNAME}`;
+  if (clientType === 'client_credentials') {
+    fenceCmd = `${fenceCmd} client-create --client ${clientName} --grant-types client_credentials`;
+  } else if (clientType === 'implicit') {
+    fenceCmd = `${fenceCmd} client-create --client ${clientName} --user ${userName} --urls https://${process.env.HOSTNAME} --grant-types implicit --public`;
+  }
 
-  if (clientType === 'implicit') {
-    fenceCmd = `${fenceCmd} --grant-types implicit --public`;
-  }
-  // see test_setup.js for the ARBORIST_... global flag setup
-  if (arboristPolicies && process.env.ARBORIST_CLIENT_POLICIES) {
-    fenceCmd = `${fenceCmd} --policies ${arboristPolicies}`;
-  }
   console.log(`running: ${fenceCmd}`);
   const resCmd = bash.runCommand(fenceCmd, 'fence', takeLastLine);
   const arr = resCmd.replace(/[()']/g, '').split(',').map((val) => val.trim());
@@ -44,7 +41,9 @@ function createClient(clientName, userName, clientType, arboristPolicies = null)
  * @param {string} clientName - client name
  */
 function deleteClient(clientName) {
-  bash.runCommand(`fence-create client-delete --client ${clientName}`, 'fence', takeLastLine);
+  const deleteClientCmd = `fence-create client-delete --client ${clientName}`;
+  const deleteClientReq = bash.runCommand(deleteClientCmd, 'fence', takeLastLine);
+  console.log(`Client deleted : ${deleteClientReq}`);
 }
 
 /**
@@ -60,6 +59,9 @@ class Client {
     this.arboristPolicies = arboristPolicies;
     this._client = null;
   }
+
+  // deleteClient = fenceTasks.deleteClient;
+  // createClient = fenceTasks.createClient;
 
   get client() {
     if (!this._client) {
@@ -87,6 +89,7 @@ module.exports = {
   /**
    * Fence endpoints
    */
+  Client,
   endpoints: {
     root: rootEndpoint,
     version: `${rootEndpoint}/_version`,
